Rename ProjectConfigurationUtils to ProjectConfigurationActions
