+++ subprojects/platform-base/src/test/groovy/dev/nokee/platform/base/internal/dependencies/ProjectConfigurationActions_ArtifactIfExistsTest.java
@@ -15,14 +15,14 @@ import java.nio.file.Path;
 
 import static dev.nokee.internal.testing.utils.ConfigurationTestUtils.testConfiguration;
 import static dev.nokee.internal.testing.utils.TestUtils.objectFactory;
-import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationUtils.*;
+import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationActions.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
 import static org.hamcrest.io.FileMatchers.aFileNamed;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-@Subject(ProjectConfigurationUtils.class)
-class ProjectConfigurationUtils_ArtifactIfExistsTest {
+@Subject(ProjectConfigurationActions.class)
+class ProjectConfigurationActions_ArtifactIfExistsTest {
 	@TempDir Path testDirectory;
 
 	@Test
