+++ subprojects/platform-base/src/test/groovy/dev/nokee/platform/base/internal/dependencies/ProjectConfigurationActions_AsConsumableTest.java
@@ -9,41 +9,41 @@ import spock.lang.Subject;
 import java.util.function.Consumer;
 
 import static dev.nokee.internal.testing.utils.ConfigurationTestUtils.testConfiguration;
-import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationUtils.*;
+import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationActions.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasToString;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-@Subject(ProjectConfigurationUtils.class)
-class ProjectConfigurationUtils_AsDeclarableTest {
+@Subject(ProjectConfigurationActions.class)
+class ProjectConfigurationActions_AsConsumableTest {
 	@Test
-	void canConfigureConfigurationAsDeclarableBucket()  {
-		val configuration = testConfiguration(asDeclarable());
-		assertThat("should not be consumable", configuration.isCanBeConsumed(), equalTo(false));
+	void canConfigureConfigurationAsConsumableBucket()  {
+		val configuration = testConfiguration(asConsumable());
+		assertThat("should be consumable", configuration.isCanBeConsumed(), equalTo(true));
 		assertThat("should not be resolvable", configuration.isCanBeResolved(), equalTo(false));
 	}
 
 	@Test
-	void canCheckWhenConfigurationIsNotDeclarable() {
+	void canCheckWhenConfigurationIsNotConsumable() {
 		val ex = assertThrows(IllegalStateException.class,
-			() -> assertConfigured(testConfiguration(), asDeclarable()));
-		assertThat(ex.getMessage(), equalTo("Cannot reuse existing configuration named 'test' as a declarable configuration because it does not match the expected configuration (expecting: [canBeConsumed: false, canBeResolved: false], actual: [canBeConsumed: true, canBeResolved: true])."));
+			() -> assertConfigured(testConfiguration(), asConsumable()));
+		assertThat(ex.getMessage(), equalTo("Cannot reuse existing configuration named 'test' as a consumable configuration because it does not match the expected configuration (expecting: [canBeConsumed: true, canBeResolved: false], actual: [canBeConsumed: true, canBeResolved: true])."));
 	}
 
 	@Test
 	@SuppressWarnings("UnstableApiUsage")
 	void checkEquals() {
 		new EqualsTester()
-			.addEqualityGroup(asDeclarable(), asDeclarable())
+			.addEqualityGroup(asConsumable(), asConsumable())
 			.addEqualityGroup(asResolvable())
-			.addEqualityGroup(asConsumable())
+			.addEqualityGroup(asDeclarable())
 			.addEqualityGroup((Consumer<Configuration>) it -> {})
 			.testEquals();
 	}
 
 	@Test
 	void checkToString() {
-		assertThat(asDeclarable(), hasToString("ProjectConfigurationUtils.asDeclarable()"));
+		assertThat(asConsumable(), hasToString("ProjectConfigurationUtils.asConsumable()"));
 	}
 }
