+++ subprojects/platform-base/src/test/groovy/dev/nokee/platform/base/internal/dependencies/ProjectConfigurationActions_AsResolvableTest.java
@@ -9,14 +9,14 @@ import spock.lang.Subject;
 import java.util.function.Consumer;
 
 import static dev.nokee.internal.testing.utils.ConfigurationTestUtils.testConfiguration;
-import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationUtils.*;
+import static dev.nokee.platform.base.internal.dependencies.ProjectConfigurationActions.*;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasToString;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-@Subject(ProjectConfigurationUtils.class)
-class ProjectConfigurationUtils_AsResolvableTest {
+@Subject(ProjectConfigurationActions.class)
+class ProjectConfigurationActions_AsResolvableTest {
 	@Test
 	void canConfigureConfigurationAsResolvableBucket()  {
 		val configuration = testConfiguration(asResolvable());
